requirements.txt
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
nbformat>=5.10.0
playwright>=1.45.0
Pillow>=10.0.0
filelock>=3.15.0
